+++ app/assets/javascripts/concepts.js
@@ -100,7 +100,7 @@ var ConceptsModule = (function () {
 
                 onSuccess();
 
-                WindowManager.loadViewerData(data, viewerID, "concept", windowType);
+                WindowManager.callLoadViewerData(data, viewerID, "concept", windowType);
 
                 // only resolve waiting requests if this is an inline viewer. New and popup viewers still have more processing.
                 if (windowType != WindowManager.NEW && windowType != WindowManager.POPUP) {
@@ -129,7 +129,7 @@ var ConceptsModule = (function () {
     return {
         initialize: init,
         createViewer: createViewer,
-        loadViewerData: loadViewerData,
+        callLoadViewerData: callLoadViewerData,
         VIEW: VIEW,
         EDIT: EDIT,
         CREATE: CREATE,